+++ src/compressed_io.h
@@ -44,8 +44,8 @@ public:
     virtual void WriteFrame(std::vector<char> data) = 0;
     virtual std::unique_ptr<TFrame> ReadFrame() = 0;
     virtual std::string GetName() const = 0;
-    virtual int GetChannelNum() const = 0;
-    virtual long long GetLengthInSamples() const = 0;
+    virtual uint8_t GetChannelNum() const = 0;
+    virtual uint64_t GetLengthInSamples() const = 0;
     virtual ~ICompressedIO() {}
 };
